+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Note that the current nightly _requires_ VS Code Insiders.
 - fix: fall back to any installed browser version if stable is not available ([ref](https://github.com/microsoft/vscode/issues/129013))
 - fix: workspaceFolder error in workspace launch configs ([ref](https://github.com/microsoft/vscode/issues/128922))
 - fix: console logs being slow when run without debugging ([#1068](https://github.com/microsoft/vscode-js-debug/issues/1068))
+- fix: not pausing on unhandled promise rejections ([ref](https://github.com/microsoft/vscode/issues/130265))
 
 ## v1.60 (August 2021)